Added support for snowflake &  optimlized mailing system

+++ Authorization/jwtManager.go
@@ -48,7 +48,7 @@ func getPublicKey() (ed25519.PublicKey, error) {
 	return ed25519.PublicKey(keyBytes), nil
 }
 
-func GenerateJWTToken(userID int32, email string, firstName string, lastName string, access string, expiration time.Time, nonce string, sessionID string, country string) (string, error) {
+func GenerateJWTToken(userID int64, email string, firstName string, lastName string, access string, expiration time.Time, nonce string, sessionID string, country string) (string, error) {
 	tokenID, err := GenerateToken()
 	if err != nil {
 		return "", err
